original agent with base agent builder structure setup
