+++ src/crewai/tools/agent_tools.py
@@ -4,7 +4,7 @@ from langchain.tools import StructuredTool
 from pydantic import BaseModel, Field
 
 
-from crewai.agents.third_party_agents.base_agent import BaseAgent
+from crewai.agents.agent_builder.base_agent import BaseAgent
 from crewai.task import Task
 from crewai.utilities import I18N
