+++ src/crewai/utilities/converter.py
@@ -3,7 +3,9 @@ import json
 from langchain.schema import HumanMessage, SystemMessage
 from langchain_openai import ChatOpenAI
 from pydantic import model_validator
-from crewai.agents.third_party_agents.utilities.converter_base import OutputConverter
+from crewai.agents.agent_builder.utilities.base_output_converter_base import (
+    OutputConverter,
+)
 
 
 class ConverterError(Exception):
